Sitecoreified the Hub Bar component
