+++ src/items/content/XmCloudPreview/home/Page Components/Default Hub Bar.yml
@@ -13,14 +13,23 @@ Languages:
       Value: 20220519T113241Z
     - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
       Hint: __Revision
-      Value: "7ce753ca-645e-42d3-a55d-08371febedec"
+      Value: "9cfda789-7a13-415c-be6b-eeeb735969ff"
     - ID: "b6c6b30a-0c42-4e3e-b47a-1a4fb37c8bc8"
       Hint: Links
-      Value: "{3860BD73-C6E3-447C-A6BD-F292857CCDE3}"
+      Value: |
+        {3860BD73-C6E3-447C-A6BD-F292857CCDE3}
+        {5198EA3E-8186-47E9-8C81-D27219B5269E}
+        {707CBBCF-E7C3-4A92-931B-43C44D2BB821}
+        {5C6FBAE1-4602-4284-B9A0-0B818DF18EE9}
+        {6ACD8B58-1A69-4FFD-B4BF-E3C112B11999}
+        {A73458FA-093A-4C33-9D9D-8443F0E90CCB}
+        {F343437E-A83A-4C44-8982-3ADD8FA46AE6}
+        {15FF6A9F-4D7D-483C-A73B-A02951CE729B}
+        {03AE15B4-6929-41E9-8AFB-EA1BCC979B8B}
     - ID: "badd9cf9-53e0-4d0c-bcc0-2d784c282f6a"
       Hint: __Updated by
       Value: |
         sitecore\matthew.peel@jet2.com
     - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
       Hint: __Updated
-      Value: 20220519T113306Z
+      Value: 20220519T143954Z
